requirements.txt
google-generativeai
pillow
numpy
scipy
python-dotenv
rich
gradio
chromadb
pillow-lut